backend/requirements.txt
django ==3.2
django-debug-toolbar==3.4.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
requests==2.28.1
django-cors-header==3.13.0